refactor: rename speakers file type to d.ts
